# app/api/openapi.yaml
openapi: 3.0.0
info:
  title: AWS Services API
  version: 1.0.0
  description: Comprehensive API for managing AWS services including Lambda, API Gateway, IAM, SNS, SQS, and Step Functions
servers:
  - url: /api
    description: API v1
components:
  securitySchemes:
    AWS4-HMAC-SHA256:
      type: apiKey
      in: header
      name: Authorization
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
        timestamp:
          type: string

    # S3 Schemas
...
    LambdaFunction:
      type: object
      properties:
        FunctionName:
          type: string
        Runtime:
          type: string
          enum: [nodejs18.x, python3.9, java11, dotnet6]
        Handler:
          type: string
        Role:
          type: string
        Code:
          type: object
          properties:
            ZipFile:
              type: string
            S3Bucket:
              type: string
            S3Key:
              type: string
        Environment:
          type: object
          properties:
            Variables:
              type: object
              additionalProperties:
                type: string
        MemorySize:
          type: integer
          minimum: 128
          maximum: 10240
        Timeout:
          type: integer
          minimum: 1
          maximum: 900
        Tags:
          type: object
          additionalProperties:
            type: string

    # API Gateway Schemas
...
    IAMRole:
      type: object
      properties:
        RoleName:
          type: string
        Description:
          type: string
        AssumeRolePolicyDocument:
          type: object
        MaxSessionDuration:
          type: integer
        Path:
          type: string
        Tags:
          type: array
          items:
            type: object
            properties:
              Key:
                type: string
              Value:
                type: string
        ManagedPolicyArns:
          type: array
          items:
            type: string
        InlinePolicies:
          type: object
          additionalProperties:
            type: object

    # SNS Schemas
...
paths:
...
  /lambda/functions:
    get:
      summary: List Lambda functions
      operationId: listFunctions
      responses:
        '200':
          description: Successfully retrieved Lambda functions
          content:
            application/json:
              schema:
                type: object
                properties:
                  functions:
                    type: array
                    items:
                      $ref: '#/components/schemas/LambdaFunction'
                  requestId:
                    type: string
                  timestamp:
                    type: string
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create Lambda function
      operationId: createFunction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LambdaFunction'
      responses:
        '201':
          description: Function created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LambdaFunction'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /iam/roles:
    get:
      summary: List IAM roles
      operationId: listRoles
      responses:
        '200':
          description: Successfully retrieved roles
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items:
                      $ref: '#/components/schemas/IAMRole'
                  requestId:
                    type: string
                  timestamp:
                    type: string
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create IAM role
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IAMRole'
      responses:
        '201':
          description: Role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IAMRole'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...